Sql Query Generator/requirements.txt
# libraries
streamlit
google-generativeai
python-dotenv
